🚚 [RUM-250] move startDeflateWorker in a new folder
