+++ packages/rum/src/domain/deflate/deflateWorker.spec.ts
@@ -4,8 +4,8 @@ import type { RumConfiguration } from '@datadog/browser-rum-core'
 import type { Clock } from '@datadog/browser-core/test'
 import { mockClock } from '@datadog/browser-core/test'
 import { MockWorker } from '../../../test'
-import type { createDeflateWorker } from './startDeflateWorker'
-import { startDeflateWorker, resetDeflateWorkerState, INITIALIZATION_TIME_OUT_DELAY } from './startDeflateWorker'
+import type { createDeflateWorker } from './deflateWorker'
+import { startDeflateWorker, resetDeflateWorkerState, INITIALIZATION_TIME_OUT_DELAY } from './deflateWorker'
 
 // Arbitrary stream ids used for tests
 const TEST_STREAM_ID = 5
