+++ packages/rum/src/domain/deflate/deflateWorker.ts
