+++ packages/rum/src/domain/deflate/index.ts
@@ -1,0 +1,1 @@
+export { DeflateWorker, startDeflateWorker } from './deflateWorker'
