+++ packages/rum/src/domain/segmentCollection/index.ts
@@ -1,4 +1,3 @@
 export { startSegmentCollection, setSegmentBytesLimit } from './segmentCollection'
-export { startDeflateWorker, DeflateWorker } from './startDeflateWorker'
 export { SEGMENT_BYTES_LIMIT } from './segmentCollection'
 export type { BrowserSegmentMetadataAndSegmentSizes } from './buildReplayPayload'
